requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
decorator==5.1.1
Django==4.2.5
django-rest-knox==4.2.0
django-user-agents==0.4.0
djangorestframework==3.14.0
dynaconf==3.2.3
future==0.18.3
geocoder==1.38.1
hvac==1.2.1
idna==3.4
knox==0.1.14
loguru==0.7.2
Pillow==10.0.0
pycparser==2.21
pyhcl==0.4.5
pytz==2023.3.post1
ratelim==0.1.6
requests==2.31.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
ua-parser==0.18.0
urllib3==2.0.4
user-agents==2.2.0
validators==0.22.0
win32-setctime==1.1.0
